a2a_basic/pyproject.toml
[project]
name = "a2a-basic"
version = "0.1.0"
description = "Simple agent to agent communication that returns a default message."
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "a2a-sdk>=0.2.5",
    "uvicorn>=0.34.3",
]

[tool.hatch.build.targets.wheel]
packages = ["."]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "ruff>=0.11.13",
]
